Warnings fixes

+++ src/xrCore/_bitwise.h
@@ -90,7 +90,7 @@ IC u64 btwCount1(u64 v)
 
 IC int iFloor(float x)
 {
-    return std::floor(x);
+    return static_cast<int>(std::floor(x));
 }
 
 /* intCeil() is a non-interesting variant, since effectively
@@ -98,7 +98,7 @@ IC int iFloor(float x)
  */
 IC int iCeil(float x)
 {
-    return std::ceil(x);
+    return static_cast<int>(std::ceil(x));
 }
 
 // Validity checks
